drivers: consolidate WS channel creation code
