+++ lib/drivers/direct/index.js
@@ -6,6 +6,7 @@ var Container = require('./container');
 var EventEmitter = require('events').EventEmitter;
 var assert = require('assert');
 var async = require('async');
+var control = require('../common/control');
 var debug = require('debug')('strong-pm:direct:driver');
 var mandatory = require('./../../util').mandatory;
 var path = require('path');
@@ -145,21 +146,14 @@ Driver.prototype._containerById = function(instanceId) {
 
   if (!container) {
     debug('create container for instance %j', instanceId);
-    var port = mandatory(this._server.port());
-    var channel = this._wsRouter.createChannel(onRequest);
-    var token = channel.getToken();
-    var control = util.format('ws://%s@127.0.0.1:%d%s',
-                              token, port, this._wsRouter.path);
-
-    debug('container control socket: %s', control);
-
+    var channel = control.create(this._server, this._wsRouter, onRequest);
     container = this._containers[instanceId] = this._Container({
       baseDir: this._baseDir,
       console: this._console,
       server: this._server,
       instanceId: instanceId,
       ctl: channel,
-      control: control,
+      control: channel.url,
     });
     container.channel = channel;
     container.instance = {
