+++ lib/drivers/docker/index.js
@@ -2,6 +2,7 @@
 
 var _ = require('lodash');
 var async = require('async');
+var control = require('../common/control');
 var debug = require('debug')('strong-pm:docker:driver');
 var Docker = require('dockerode');
 var EventEmitter = require('events').EventEmitter;
@@ -11,7 +12,6 @@ var inherits = require('util').inherits;
 var isError = require('util').isError;
 var logBuffer = require('../common/log-buffer');
 var mandatory = require('./../../util').mandatory;
-var os = require('os');
 var path = require('path');
 
 module.exports = exports = DockerDriver;
@@ -386,20 +386,14 @@ DockerDriver.prototype._instance = function(id) {
       driverMeta: {},
       svcDir: path.resolve(this.baseDir, 'svc', String(id)),
     });
-    var port = mandatory(this.server.port());
-    var channel = this._wsRouter.createChannel(onInstanceRequest);
-    var token = channel.getToken();
-    var addr = publicIp();
-    var wsPath = this._wsRouter.path;
-    var control = fmt('ws://%s@%s:%d%s', token, addr, port, wsPath);
-    debug('container control socket: %s', control);
+    var channel = control.create(this.server, this._wsRouter, onRequest);
     instance.channel = channel;
-    instance.startOpts.control = control;
+    instance.startOpts.control = channel.url;
     this._instances[id] = instance;
   }
   return this._instances[id];
 
-  function onInstanceRequest(req) {
+  function onRequest(req) {
     debug('Request via ws(%s):', id, req);
     instance.emit('request', req);
   }
@@ -412,14 +406,3 @@ DockerDriver.prototype.getName = function() {
 DockerDriver.prototype.getStatus = function() {
   return 'running';
 };
-
-// TODO:
-// This is a simple stand-in to get things going. Realistically, we'll need to
-// provide some way for the user to configure what host should be given out for
-// use in URLs back to us, since we could be behind a proxy or load balancer or
-// any number of odd network layers.
-function publicIp() {
-  var eths = os.networkInterfaces();
-  var addrs = eths.en0 || eths.eth0;
-  return _(addrs).where({family: 'IPv4'}).pluck('address').first();
-}
